move MySQL-specific closure-table tests to a separate file
